Move RefreshToken to entity package
