+++ src/main/java/com/example/ottokeng/domain/oauth/repository/RefreshTokenRepository.java
@@ -1,9 +1,7 @@
 package com.example.ottokeng.domain.oauth.repository;
 
-import com.example.ottokeng.domain.oauth.dto.RefreshToken;
+import com.example.ottokeng.domain.oauth.entity.RefreshToken;
 import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
 
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
 }
