+++ src/main/java/com/example/ottokeng/domain/oauth/service/OauthService.java
@@ -4,7 +4,7 @@ import com.example.ottokeng.domain.oauth.config.OauthAttributes;
 import com.example.ottokeng.domain.oauth.config.OauthProvider;
 import com.example.ottokeng.domain.oauth.dto.LoginResponse;
 import com.example.ottokeng.domain.oauth.dto.OauthTokenResponse;
-import com.example.ottokeng.domain.oauth.dto.RefreshToken;
+import com.example.ottokeng.domain.oauth.entity.RefreshToken;
 import com.example.ottokeng.domain.oauth.dto.UserProfile;
 import com.example.ottokeng.domain.oauth.repository.InMemoryProviderRepository;
 import com.example.ottokeng.domain.oauth.repository.RefreshTokenRepository;
