--- src/main/java/com/example/ottokeng/domain/user/service/UserService.java
@@ -1,4 +1,0 @@
-package com.example.ottokeng.domain.user.service;
-
-public class UserService {
-}
